+++ src/main/java/org/example/fashion_api/Models/Products/Product.java
@@ -25,9 +25,9 @@ public class Product {
 
     private String productName;
 
-    private Integer price;
+    private Long price;
 
-    private Integer discountPrice;
+    private Long discountPrice;
 
     private Integer discountPercent;
