+++ src/main/java/org/example/fashion_api/Models/Products/UpdateProductDto.java
@@ -16,9 +16,9 @@ public class UpdateProductDto  {
     private String productName;
 
     @NotNull
-    private Integer price;
+    private Long price;
 
-    private Integer discountPrice;
+    private Long discountPrice;
 
     private String brand;
